iOS: move gestures.rb to device/gestures
