+++ lib/calabash/ios.rb
@@ -20,7 +20,7 @@ module Calabash
     require 'calabash/ios/device/routes/map_route'
     require 'calabash/ios/device/routes/uia_route'
 
-    require 'calabash/ios/gestures'
+    require 'calabash/ios/device/gestures'
 
     require 'calabash/ios/status_bar'
