+++ lib/calabash/ios/device/gestures.rb
