+++ spec/lib/ios/device/gestures.rb
@@ -1,0 +1,3 @@
+describe Calabash::IOS::Gestures do
+
+end
